+++ rio/component_meta.py
@@ -48,10 +48,6 @@ class ComponentMeta(RioDataclassMeta):
     # or from a library.
     _rio_builtin_: bool
 
-    # A dict mapping old, deprecated, parameter names to the new one. This dict
-    # is modified by the `@deprecations.parameter_renamed` decorator.
-    _deprecated_parameter_names_: dict[str, str]
-
     def __init__(cls, *args, **kwargs):
         # Is this class built into Rio?
         cls._rio_builtin_ = cls.__module__.startswith("rio.")
@@ -105,9 +101,6 @@ class ComponentMeta(RioDataclassMeta):
                 for arg in args:
                     cls._rio_event_handlers_[event_tag].append((member, arg))
 
-        # Initialize class variables
-        cls._deprecated_parameter_names_ = {}
-
     def _initialize_state_properties(cls) -> None:
         """
         Spawn `StateProperty` instances for all annotated properties in this
@@ -155,9 +148,6 @@ class ComponentMeta(RioDataclassMeta):
             )
 
         # Remap deprecated parameter names to new ones
-        deprecations._remap_kwargs(
-            cls.__name__, kwargs, cls._deprecated_parameter_names_
-        )
         deprecations.remap_width_and_height(kwargs)
 
         component: C = object.__new__(cls)